refactor(manager): Move manager routes under /manager, add public root.
A1 of the app merge. The RR8 app gains a public root: root.tsx is now
a chrome-less document (html/head, color-scheme attribute, error
boundary) and everything it did before — shell, sidebar, auth and
championship middleware — moves to a new manager-layout route nested
under the /manager path prefix.

Drops basename and the assetsDir hack, both of which only existed to
fake the prefix from outside. A placeholder page holds "/" until the
public routes are ported.
