+++ apps/manager/react-router.config.ts
@@ -1,6 +1,5 @@
 import type { Config } from "@react-router/dev/config";
 
 export default {
-  basename: "/manager",
   buildDirectory: "dist",
 } satisfies Config;
