+++ apps/manager/vite.config.ts
@@ -4,6 +4,5 @@ import tailwindcss from "@tailwindcss/vite";
 import { type UserConfig } from "vite";
 
 export default {
-  build: { assetsDir: "manager/assets" },
   plugins: [cloudflare({ viteEnvironment: { name: "ssr" } }), reactRouter(), tailwindcss()],
 } satisfies UserConfig;
